Replaced occurances of 'stoi' with 'stoull', and 'int' with 'unsigned int' to reach maximum integer values of 4294967295.

+++ src/Compiler.cpp
@@ -568,7 +568,7 @@ unsigned int Compiler::evaluateBracket(std::string originalLine)
             }
             else //Else if number
             {
-                igen.genCreateInt("", stoi(data));
+                igen.genCreateInt("", stoull(data));
             }
         }
     };
